+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add path to node and scope. Allow to specify a scope in supplier keys
 - Add built method to ScopeBluePrint
 - Add Scope.mockContent
+- Add Scope.findScope
 
 ### Changed